move the connect directory from the simulation directory to the building directory
